requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
altair==5.2.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==23.2.0
babel==2.17.0
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.3.2
certifi==2025.1.31
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
configobj==5.0.8
configparser==6.0.1
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
dash==4.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.13.1
fitz==0.0.1.dev2
Flask==2.3.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.47.2
fqdn==1.5.1
fsspec==2025.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
google-pasta==0.2.0
googleapis-common-protos==1.73.1
greenlet==3.1.1
grpcio==1.73.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib-metadata==7.0.1
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.5
libclang==18.1.1
looseversion==1.3.0
lxml==5.1.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.1.0
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.3
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.92
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdfkit==1.0.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.0.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.27.2
protobuf==5.29.6
prov==2.0.0
psutil==7.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycaw==20240210
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.19.2
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.2
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==7.0.0
referencing==0.33.0
regex==2025.10.23
reportlab==4.4.10
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.17.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.3.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
simplejson==3.19.2
simsimd==6.5.16
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.47.3
streamlit==1.46.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==8.2.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.1.0
timm==1.0.26
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.22.1
toml==0.10.2
toolz==0.12.1
torch==2.11.0
torchvision==0.26.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
traits==6.3.2
transformers==4.57.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
WTForms==3.2.1
yarg==0.1.9
zipp==3.21.0
